+++ packages/components/src/font-size-picker/index.tsx
@@ -7,7 +7,7 @@ import type { ForwardedRef } from 'react';
  * WordPress dependencies
  */
 import deprecated from '@wordpress/deprecated';
-import { __, sprintf } from '@wordpress/i18n';
+import { __ } from '@wordpress/i18n';
 import { settings } from '@wordpress/icons';
 import { useState, useMemo, forwardRef } from '@wordpress/element';
 
@@ -22,24 +22,10 @@ import {
 	parseQuantityAndUnitFromRawValue,
 	useCustomUnits,
 } from '../unit-control';
-import CustomSelectControl from '../custom-select-control';
 import { VisuallyHidden } from '../visually-hidden';
-import {
-	ToggleGroupControl,
-	ToggleGroupControlOption,
-} from '../toggle-group-control';
-import {
-	getFontSizeOptions,
-	getSelectedOption,
-	isSimpleCssValue,
-	CUSTOM_FONT_SIZE,
-} from './utils';
+import { isSimpleCssValue } from './utils';
 import { HStack } from '../h-stack';
-import type {
-	FontSizePickerProps,
-	FontSizeSelectOption,
-	FontSizeToggleGroupOption,
-} from './types';
+import type { FontSizePickerProps } from './types';
 import {
 	Container,
 	HeaderHint,
@@ -48,6 +34,8 @@ import {
 	ResetButton,
 } from './styles';
 import { Spacer } from '../spacer';
+import FontSizePickerSelect from './font-size-picker-select';
+import FontSizePickerToggleGroup from './font-size-picker-toggle-group';
 
 const UnforwardedFontSizePicker = (
 	props: FontSizePickerProps,
@@ -65,6 +53,7 @@ const UnforwardedFontSizePicker = (
 		withSlider = false,
 		withReset = true,
 	} = props;
+
 	if ( ! __nextHasNoMarginBottom ) {
 		deprecated( 'Bottom margin styles for wp.components.FontSizePicker', {
 			since: '6.1',
@@ -77,28 +66,16 @@ const UnforwardedFontSizePicker = (
 		availableUnits: [ 'px', 'em', 'rem' ],
 	} );
 
-	/**
-	 * The main font size UI displays a toggle group when the presets are less
-	 * than six and a select control when they are more.
-	 */
-	const fontSizesContainComplexValues = fontSizes.some(
-		( { size: sizeArg } ) => ! isSimpleCssValue( sizeArg )
-	);
 	const shouldUseSelectControl = fontSizes.length > 5;
-	const options = useMemo(
-		() =>
-			getFontSizeOptions(
-				shouldUseSelectControl,
-				fontSizes,
-				disableCustomFontSizes
-			),
-		[ shouldUseSelectControl, fontSizes, disableCustomFontSizes ]
+	const selectedFontSize = fontSizes.find(
+		( fontSize ) => fontSize.size === value
 	);
-	const selectedOption = getSelectedOption( fontSizes, value );
-	const isCustomValue = selectedOption.slug === CUSTOM_FONT_SIZE;
+	const isCustomValue = !! value && ! selectedFontSize;
+
 	const [ showCustomValueControl, setShowCustomValueControl ] = useState(
 		! disableCustomFontSizes && isCustomValue
 	);
+
 	const headerHint = useMemo( () => {
 		if ( showCustomValueControl ) {
 			return `(${ __( 'Custom' ) })`;
@@ -113,22 +90,30 @@ const UnforwardedFontSizePicker = (
 				`(${ value })`
 			);
 		}
+
 		if ( shouldUseSelectControl ) {
 			return (
-				selectedOption?.size !== undefined &&
-				isSimpleCssValue( selectedOption?.size ) &&
-				`(${ selectedOption?.size })`
+				selectedFontSize?.size !== undefined &&
+				isSimpleCssValue( selectedFontSize?.size ) &&
+				`(${ selectedFontSize?.size })`
 			);
 		}
 
+		if ( ! selectedFontSize ) {
+			return __( 'Default' );
+		}
+
 		// Calculate the `hint` for toggle group control.
-		let hint = selectedOption?.name || selectedOption.slug;
+		let hint = selectedFontSize.name || selectedFontSize.slug;
+		const fontSizesContainComplexValues = fontSizes.some(
+			( fontSize ) => ! isSimpleCssValue( fontSize.size )
+		);
 		if (
 			! fontSizesContainComplexValues &&
-			typeof selectedOption.size === 'string'
+			typeof selectedFontSize.size === 'string'
 		) {
 			const [ , unit ] = parseQuantityAndUnitFromRawValue(
-				selectedOption.size,
+				selectedFontSize.size,
 				units
 			);
 			hint += `(${ unit })`;
@@ -136,25 +121,16 @@ const UnforwardedFontSizePicker = (
 		return hint;
 	}, [
 		showCustomValueControl,
-		selectedOption?.name,
-		selectedOption?.size,
-		value,
 		isCustomValue,
+		selectedFontSize,
+		value,
 		shouldUseSelectControl,
-		fontSizesContainComplexValues,
+		fontSizes,
 	] );
 
-	if ( ! options ) {
+	if ( fontSizes.length === 0 && disableCustomFontSizes ) {
 		return null;
 	}
-
-	// This is used for select control only. We need to add support
-	// for ToggleGroupControl.
-	const currentFontSizeSR = sprintf(
-		// translators: %s: Currently selected font size.
-		__( 'Currently selected font size: %s' ),
-		selectedOption.name
-	);
 
 	// If neither the value or first font size is a string, then FontSizePicker
 	// operates in a legacy "unitless" mode where UnitControl can only be used
@@ -210,43 +186,31 @@ const UnforwardedFontSizePicker = (
 				{ !! fontSizes.length &&
 					shouldUseSelectControl &&
 					! showCustomValueControl && (
-						<CustomSelectControl
-							__nextUnconstrainedWidth
-							className="components-font-size-picker__select"
-							label={ __( 'Font size' ) }
-							hideLabelFromVision
-							describedBy={ currentFontSizeSR }
-							options={ options as FontSizeSelectOption[] }
-							value={ ( options as FontSizeSelectOption[] ).find(
-								( option ) => option.key === selectedOption.slug
-							) }
-							onChange={ ( {
-								selectedItem,
-							}: {
-								selectedItem: FontSizeSelectOption;
-							} ) => {
-								if ( selectedItem.size === undefined ) {
+						<FontSizePickerSelect
+							fontSizes={ fontSizes }
+							value={ value }
+							disableCustomFontSizes={ disableCustomFontSizes }
+							size={ size }
+							onChange={ ( newValue ) => {
+								if ( newValue === undefined ) {
 									onChange?.( undefined );
 								} else {
 									onChange?.(
-										hasUnits
-											? selectedItem.size
-											: Number( selectedItem.size )
+										hasUnits ? newValue : Number( newValue )
 									);
 								}
-								if ( selectedItem.key === CUSTOM_FONT_SIZE ) {
-									setShowCustomValueControl( true );
-								}
 							} }
-							size={ size }
+							onSelectCustom={ () =>
+								setShowCustomValueControl( true )
+							}
 						/>
 					) }
 				{ ! shouldUseSelectControl && ! showCustomValueControl && (
-					<ToggleGroupControl
-						__nextHasNoMarginBottom={ __nextHasNoMarginBottom }
-						label={ __( 'Font size' ) }
-						hideLabelFromVision
+					<FontSizePickerToggleGroup
+						fontSizes={ fontSizes }
 						value={ value }
+						__nextHasNoMarginBottom={ __nextHasNoMarginBottom }
+						size={ size }
 						onChange={ ( newValue ) => {
 							if ( newValue === undefined ) {
 								onChange?.( undefined );
@@ -256,21 +220,7 @@ const UnforwardedFontSizePicker = (
 								);
 							}
 						} }
-						isBlock
-						size={ size }
-					>
-						{ ( options as FontSizeToggleGroupOption[] ).map(
-							( option ) => (
-								<ToggleGroupControlOption
-									key={ option.key }
-									value={ option.value }
-									label={ option.label }
-									aria-label={ option.name }
-									showTooltip={ true }
-								/>
-							)
-						) }
-					</ToggleGroupControl>
+					/>
 				) }
 				{ ! disableCustomFontSizes && showCustomValueControl && (
 					<Flex className="components-font-size-picker__custom-size-control">